refactor: 💡 NetworkingService and UtilityView as internal
